# packages/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FITVO API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/patients/{tenantId}/invites": {
      "post": {
        "summary": "Convida um paciente para uma especialidade (profissional)",
        "tags": [
          "patient"
        ],
        "description": "Cria um convite de uso unico direcionado a UMA especialidade (D-006/D-052), declarando a MODALIDADE do atendimento (D-101), que o vinculo herda no aceite. Requer perfil profissional no tenant que reivindica a especialidade-alvo E e-mail verificado (D-029). Devolve o token em claro UMA vez; o banco guarda apenas o hash.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "E-mail do paciente convidado."
                  },
                  "specialtyId": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Especialidade-alvo do vinculo (D-052)."
                  },
                  "modality": {
                    "type": "string",
                    "enum": [
                      "ONLINE",
                      "PRESENCIAL",
                      "HIBRIDO"
                    ],
                    "description": "Modalidade do atendimento (D-101). Definida pelo PROFISSIONAL."
                  }
                },
                "required": [
                  "email",
                  "specialtyId",
                  "modality"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "tenantId",
            "required": true,
            "description": "ID do tenant do profissional."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "invite": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "specialtyId": {
                          "type": "string"
                        },
                        "modality": {
                          "type": "string",
                          "enum": [
                            "ONLINE",
                            "PRESENCIAL",
                            "HIBRIDO"
                          ],
                          "description": "Modalidade do atendimento (D-101). Definida pelo PROFISSIONAL."
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "PENDING",
                            "ACCEPTED",
                            "EXPIRED",
                            "REVOKED"
                          ]
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "specialtyId",
                        "modality",
                        "status",
                        "expiresAt",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    },
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "invite",
                    "token"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{tenantId}/overview": {
      "get": {
        "summary": "Lista convites pendentes e vinculos ativos do profissional",
        "tags": [
          "patient"
        ],
        "description": "Visao OPERACIONAL do proprio profissional: convites pendentes + vinculos ativos (ADR-0001). Requer perfil profissional no tenant.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "tenantId",
            "required": true,
            "description": "ID do tenant do profissional."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pendingInvites": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "specialtyId": {
                            "type": "string"
                          },
                          "modality": {
                            "type": "string",
                            "enum": [
                              "ONLINE",
                              "PRESENCIAL",
                              "HIBRIDO"
                            ],
                            "description": "Modalidade do atendimento (D-101). Definida pelo PROFISSIONAL."
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "PENDING",
                              "ACCEPTED",
                              "EXPIRED",
                              "REVOKED"
                            ]
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          }
                        },
                        "required": [
                          "id",
                          "email",
                          "specialtyId",
                          "modality",
                          "status",
                          "expiresAt",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "activeBonds": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "patientProfileId": {
                            "type": "string"
                          },
                          "patientName": {
                            "type": "string"
                          },
                          "patientEmail": {
                            "type": "string"
                          },
                          "specialtyId": {
                            "type": "string"
                          },
                          "modality": {
                            "type": "string",
                            "enum": [
                              "ONLINE",
                              "PRESENCIAL",
                              "HIBRIDO"
                            ],
                            "description": "Modalidade do atendimento (D-101). Definida pelo PROFISSIONAL."
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "ACTIVE",
                              "ARCHIVED"
                            ]
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          "archivedAt": {
                            "nullable": true,
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          }
                        },
                        "required": [
                          "id",
                          "patientProfileId",
                          "patientName",
                          "patientEmail",
                          "specialtyId",
                          "modality",
                          "status",
                          "createdAt",
                          "archivedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "pendingInvites",
                    "activeBonds"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{tenantId}/invites/{inviteId}/resend": {
      "post": {
        "summary": "Reenvia um convite pendente (profissional)",
        "tags": [
          "patient"
        ],
        "description": "Rotaciona o token e estende a validade na mesma linha (D-055). Devolve o novo token em claro UMA vez. Requer perfil profissional no tenant e e-mail verificado (D-029).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "tenantId",
            "required": true,
            "description": "ID do tenant do profissional."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "inviteId",
            "required": true,
            "description": "ID do convite."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "invite": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "specialtyId": {
                          "type": "string"
                        },
                        "modality": {
                          "type": "string",
                          "enum": [
                            "ONLINE",
                            "PRESENCIAL",
                            "HIBRIDO"
                          ],
                          "description": "Modalidade do atendimento (D-101). Definida pelo PROFISSIONAL."
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "PENDING",
                            "ACCEPTED",
                            "EXPIRED",
                            "REVOKED"
                          ]
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "specialtyId",
                        "modality",
                        "status",
                        "expiresAt",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    },
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "invite",
                    "token"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{tenantId}/invites/{inviteId}/revoke": {
      "post": {
        "summary": "Revoga um convite pendente (profissional)",
        "tags": [
          "patient"
        ],
        "description": "Marca o convite como REVOKED. Requer perfil profissional no tenant.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "tenantId",
            "required": true,
            "description": "ID do tenant do profissional."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "inviteId",
            "required": true,
            "description": "ID do convite."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/patients/{tenantId}/bonds/{bondId}/archive": {
      "post": {
        "summary": "Arquiva um vinculo ativo (profissional)",
        "tags": [
          "patient"
        ],
        "description": "ACTIVE -> ARCHIVED + archivedAt. NUNCA apaga: dados preservados (D-053). Requer perfil profissional no tenant.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "tenantId",
            "required": true,
            "description": "ID do tenant do profissional."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "bondId",
            "required": true,
            "description": "ID do vinculo."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}